requirements.txt
flask
scikit-learn
spacy
pdfplumber
python-docx
pandas
sentence-transformers
torch
gunicorn

